+++ src/settings/LocationLocations/LocationDetail.js
@@ -9,9 +9,13 @@ import {
   ExpandAllButton,
   KeyValue,
   Row,
-  List
+  List,
+  Pane,
+  PaneMenu,
+  IconButton,
 } from '@folio/stripes/components';
 import { ViewMetaData } from '@folio/stripes/smart-components';
+import { stripesConnect } from '@folio/stripes/core';
 
 class LocationDetail extends React.Component {
   static manifest = Object.freeze({
@@ -40,10 +44,12 @@ class LocationDetail extends React.Component {
       libraries: PropTypes.object,
     }).isRequired,
     servicePointsById: PropTypes.object,
+    onEdit: PropTypes.func.isRequired,
+    onClose: PropTypes.func.isRequired,
   };
 
   constructor(props) {
-    super();
+    super(props);
 
     this.handleSectionToggle = this.handleSectionToggle.bind(this);
     this.handleExpandAll = this.handleExpandAll.bind(this);
@@ -138,10 +144,32 @@ class LocationDetail extends React.Component {
       );
     });
 
-
     const { sections } = this.state;
+
+    const lastMenu = (
+      <PaneMenu>
+        <FormattedMessage id="stripes-components.button.edit">
+          {ariaLabel => (
+            <IconButton
+              icon="edit"
+              onClick={() => this.props.onEdit(loc)}
+              id="clickable-edit-item"
+              ariaLabel={ariaLabel}
+              size="medium"
+            />
+          )}
+        </FormattedMessage>
+      </PaneMenu>
+    );
+
     return (
-      <div>
+      <Pane
+        paneTitle={loc.name}
+        defaultWidth="70%"
+        dismissible
+        lastMenu={lastMenu}
+        onClose={this.props.onClose}
+      >
         <Row end="xs">
           <Col xs>
             <ExpandAllButton accordionStatus={sections} onToggle={this.handleExpandAll} />
@@ -242,9 +270,9 @@ class LocationDetail extends React.Component {
         >
           {details}
         </Accordion>
-      </div>
+      </Pane>
     );
   }
 }
 
-export default LocationDetail;
+export default stripesConnect(LocationDetail);