+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.12.0 (IN PROGRESS)
 
 * Remove unnecessary permissions. Refs UITEN-7, UIORG-150.
+* Refactor location locations functionality to get rid of EntryManager usage. Refs UIORG-142.
 
 ## [2.11.0](https://github.com/folio-org/ui-organization/tree/v2.11.0) (2019-06-13)
 [Full Changelog](https://github.com/folio-org/ui-organization/compare/v2.10.0...v2.11.0)